splitting off script optimization CSE into its own source files
